backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic[dotenv]
sqlalchemy
alembic
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
requests
swarms==6.3.9
groq==0.11.0
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
PyPDF2==3.0.1
pydantic==2.5.0